LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/navigate-intercept-interrupted-expected.txt


FAIL interrupted navigate() promises with intercept() assert_array_equals: expected property 0 to be "" but got "#2" (expected array ["", "#1", "#2"] got ["#2", "#2", "#2"])
